feat: 로또 구현

+++ src/main/java/lotto/lotto/domain/LottoResult.java
@@ -36,9 +36,9 @@ public enum LottoResult {
                 .orElse(LOSE);
     }
 
-    public static int getTotalPrize(List<LottoResult> results) {
+    public static long getTotalPrize(List<LottoResult> results) {
         return results.stream()
-                .mapToInt(LottoResult::getPrize)
+                .mapToLong(LottoResult::getPrize)
                 .sum();
     }
